Added sorting to status page and moved some files
