--- src/pages/basic_data/app_admin/NviAdminStatusPage.tsx
@@ -1,96 +1,0 @@
-import {
-  Box,
-  Paper,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  Typography,
-} from '@mui/material';
-import { Trans, useTranslation } from 'react-i18next';
-import { useGetUrlFilteredInstitutionReports } from '../../../api/hooks/useGetUrlFilteredInstitutionReports';
-import { PercentageWithIcon } from '../../../components/atoms/PercentageWithIcon';
-import { TableSkeleton } from '../../../components/skeletons/TableSkeleton';
-import { HorizontalBox } from '../../../components/styled/Wrappers';
-import { InstitutionReport } from '../../../types/nvi.types';
-import { getNviInstitutionName, getNviSectorLabel } from '../../../utils/nviAdminReportSelectors';
-import { NviStatusWrapper } from '../../messages/components/NviStatusWrapper';
-
-export const NviAdminStatusPage = () => {
-  const { t } = useTranslation();
-  const { filteredData, isPending, isError } = useGetUrlFilteredInstitutionReports();
-
-  return (
-    <NviStatusWrapper
-      headline={t('basic_data.nvi.reporting_status')}
-      topView={
-        <Box sx={{ mb: '1rem' }}>
-          <Trans
-            t={t}
-            i18nKey="basic_data.nvi.reporting_status_description"
-            components={{ p: <Typography gutterBottom /> }}
-          />
-        </Box>
-      }
-      yearSelector
-      sectorSelector
-      institutionSearch>
-      {isPending ? (
-        <TableSkeleton />
-      ) : isError ? (
-        <Typography>{t('feedback.error.get_nvi_reports')}</Typography>
-      ) : (
-        <TableContainer component={Paper} variant="outlined">
-          <Table size="small">
-            <TableHead>
-              <TableRow sx={{ whiteSpace: 'nowrap', bgcolor: 'white' }}>
-                <TableCell>{t('common.institution')}</TableCell>
-                <TableCell>{t('sector')}</TableCell>
-                <TableCell>{t('candidate')}</TableCell>
-                <TableCell>{t('controlling')}</TableCell>
-                <TableCell>{t('approved')}</TableCell>
-                <TableCell>{t('rejected')}</TableCell>
-                <TableCell>{t('disputes')}</TableCell>
-                <TableCell>{t('common.total_number')}</TableCell>
-                <TableCell>{t('percentage_controlled')}</TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {filteredData.map((report: InstitutionReport) => {
-                const { id, institutionSummary } = report;
-                const { byLocalApprovalStatus, totals } = institutionSummary;
-                const percentageControlled =
-                  totals.undisputedTotalCount > 0
-                    ? (byLocalApprovalStatus.approved + byLocalApprovalStatus.rejected) / totals.undisputedTotalCount
-                    : 0;
-
-                return (
-                  <TableRow key={id} sx={{ height: '4rem' }}>
-                    <TableCell>{getNviInstitutionName(report)}</TableCell>
-                    <TableCell>{getNviSectorLabel(report, t)}</TableCell>
-                    <TableCell align="center">{byLocalApprovalStatus.new}</TableCell>
-                    <TableCell align="center">{byLocalApprovalStatus.pending}</TableCell>
-                    <TableCell align="center">{byLocalApprovalStatus.approved}</TableCell>
-                    <TableCell align="center">{byLocalApprovalStatus.rejected}</TableCell>
-                    <TableCell align="center">{totals.disputedCount}</TableCell>
-                    <TableCell align="center">{totals.undisputedTotalCount}</TableCell>
-                    <TableCell align="center">
-                      <HorizontalBox sx={{ justifyContent: 'center' }}>
-                        <PercentageWithIcon
-                          displayPercentage={Math.floor(percentageControlled * 100)}
-                          alternativeIfZero={'-'}
-                        />
-                      </HorizontalBox>
-                    </TableCell>
-                  </TableRow>
-                );
-              })}
-            </TableBody>
-          </Table>
-        </TableContainer>
-      )}
-    </NviStatusWrapper>
-  );
-};
